git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@9440 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/lps/include/mcrl2/lps/invariant_checker.h
@@ -15,7 +15,7 @@
 #define INVARIANT_CHECKER_H
 
 #include <string>
-#include "aterm2.h"
+#include "mcrl2/aterm/aterm2.h"
 #include "mcrl2/data/rewriter.h"
 #include "mcrl2/data/detail/bdd_prover.h"
 #include "mcrl2/data/detail/prover/bdd2dot.h"
@@ -30,7 +30,7 @@
 /// of an mCRL2 LPS. The constructor Invariant_Checker::Invariant_Checker initializes the BDD based prover with the
 /// parameters a_rewrite_strategy, a_time_limit, a_path_eliminator, a_solver_type and the data specification of the LPS
 /// passed as parameter a_lps. The parameter a_rewrite_strategy specifies which rewrite strategy is used by the prover's
-/// rewriter. It can be set to either GS_REWR_INNER, GS_REWR_INNERC, GS_REWR_JITTY or GS_REWR_JITTYC. The parameter
+/// rewriter. It can be set to either GS_REWR_JITTY or GS_REWR_JITTYC. The parameter
 /// a_time_limit specifies the maximum amount of time in seconds to be spent by the prover on proving a single expression.
 /// If a_time_limit is set to 0, no time limit will be enforced. The parameter a_path_eliminator specifies whether or not
 /// path elimination is applied. When path elimination is applied, the prover uses an SMT solver to remove inconsistent
@@ -82,7 +82,7 @@ class Invariant_Checker
     bool f_all_violations;
     std::string f_dot_file_name;
     void print_counter_example();
-    void save_dot_file(int a_summand_number);
+    void save_dot_file(size_t a_summand_number);
     bool check_init(const data::data_expression a_invariant);
     bool check_summand(const data::data_expression a_invariant, const action_summand a_summand, const size_t a_summand_number);
     bool check_summands(const data::data_expression a_invariant);
